refactor(plotting)
